requirements.txt
alembic==1.12.0
bcrypt==4.1.2
bidict==0.23.1
blinker==1.6.2
boto3==1.34.50
botocore==1.34.50
cached-property==1.5.2
cachelib==0.12.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.0.12
click==7.1.2
colorclass==2.2.2
contourpy==1.0.6
CoolProp==6.4.3.post1
cryptography==42.0.4
cycler==0.11.0
dnspython==2.4.2
docopt==0.6.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
Flask==1.1.4
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-MongoAlchemy==0.7.0
Flask-PyMongo==2.3.0
Flask-Script==2.0.6
Flask-SocketIO==5.3.6
Flask-Sockets==0.2.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fonttools==4.38.0
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==1.0.1
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.0.1
matplotlib==3.7.1
MongoAlchemy==0.21
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pip-upgrader==1.4.15
pipdeptree==2.15.1
pluggy==1.3.0
protobuf==4.21.12
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
pymongo==4.5.0
pyparsing==3.0.9
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.9.0
python-socketio==5.11.1
requests==2.27.1
s3transfer==0.10.0
scipy==1.11.1
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==1.4.49
sqlalchemy-orm==1.2.10
terminaltables==3.1.10
tomli==2.0.1
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.14
websockets==10.4
Werkzeug==1.0.1
wsproto==1.2.0
WTForms==3.0.1
zope.event==4.6
zope.interface==5.5.2
